refactor(ui): extract slide_list scroll + reorder helpers to keep file under 1000-line cap (#271)
CI Quality Checks failed: slide_list.rs grew to 1082 lines (>1000 hard
cap) after Tasks 2-3 added the lookahead/wheel/load-at-start scroll
helpers. Extracted to:

- slide_list_scroll.rs (new): COLUMNS_PER_ROW, DEFAULT_WHEEL_STEP_PX,
  scroll_slide_into_view, scroll_slides_to_top, step_for_wheel, and
  handle_wheel_event (the wheel handler body, lifted out of the inline
  on:wheel closure).
- slide_list_utils.rs: reorder_slide_ids + its 8 unit tests (renamed to
  reorder_*).

slide_list.rs is now 922 lines. No behavior change; pure mechanical
refactor.
