+++ crates/presenter-ui/src/components/slide_list_scroll.rs
@@ -1,0 +1,146 @@
+//! Operator slide-list scroll behavior. Extracted from `slide_list.rs` to
+//! keep that file under the project's 1000-line cap.
+//!
+//! Issue #271: lookahead scroll, deterministic wheel scroll, scroll-to-top
+//! on song open.
+
+use wasm_bindgen::JsCast;
+
+/// Number of columns in the `.operator__slides` grid (CSS:
+/// `grid-template-columns: repeat(3, minmax(0, 1fr))`). The next-row anchor
+/// for an active slide at DOM index N is the slide at index N + COLUMNS_PER_ROW.
+const COLUMNS_PER_ROW: usize = 3;
+
+/// Default fallback step for wheel scroll (pixels) when no slide card is
+/// rendered yet to measure.
+const DEFAULT_WHEEL_STEP_PX: f64 = 120.0;
+
+/// Lookahead-aware scroll: ensures the active slide AND the next row of
+/// slides are visible in the `.operator__slides` container. If the active
+/// slide is on the last row (no next-row anchor), falls back to "ensure
+/// active is visible". If the active slide is above the viewport (backward
+/// navigation), top-aligns it.
+///
+/// Issue #271 concern 1.
+pub(super) fn scroll_slide_into_view(slide_id: &str) {
+    let Some(document) = web_sys::window().and_then(|w| w.document()) else {
+        return;
+    };
+    let active_selector = format!(".operator__slides [data-slide-id=\"{slide_id}\"]");
+    let Ok(Some(active_el)) = document.query_selector(&active_selector) else {
+        return;
+    };
+    let Ok(Some(container_el)) = active_el.closest(".operator__slides") else {
+        return;
+    };
+    let Ok(container) = container_el.dyn_into::<web_sys::HtmlElement>() else {
+        return;
+    };
+    let Ok(active_html) = active_el.dyn_into::<web_sys::HtmlElement>() else {
+        return;
+    };
+
+    let container_rect = container.get_bounding_client_rect();
+    let active_rect = active_html.get_bounding_client_rect();
+    let scroll_top = container.scroll_top() as f64;
+
+    // Backward navigation: top-align the active slide if it's above the viewport.
+    if active_rect.top() < container_rect.top() {
+        let delta = container_rect.top() - active_rect.top();
+        container.set_scroll_top((scroll_top - delta) as i32);
+        return;
+    }
+
+    // Find the next-row anchor: the slide at active_index + COLUMNS_PER_ROW
+    // in DOM order within the same container.
+    let cards = container.query_selector_all("[data-slide-id]").ok();
+    let next_row_el: Option<web_sys::HtmlElement> = cards.and_then(|nodes| {
+        let mut active_index: Option<usize> = None;
+        for i in 0..nodes.length() {
+            if let Some(node) = nodes.item(i) {
+                if let Ok(el) = node.dyn_into::<web_sys::Element>() {
+                    if el.get_attribute("data-slide-id").as_deref() == Some(slide_id) {
+                        active_index = Some(i as usize);
+                        break;
+                    }
+                }
+            }
+        }
+        let target_index = active_index? + COLUMNS_PER_ROW;
+        nodes
+            .item(target_index as u32)
+            .and_then(|n| n.dyn_into::<web_sys::HtmlElement>().ok())
+    });
+
+    if let Some(anchor) = next_row_el {
+        // Scroll so the next-row anchor's bottom is at the container's bottom.
+        let anchor_rect = anchor.get_bounding_client_rect();
+        if anchor_rect.bottom() > container_rect.bottom() {
+            let delta = anchor_rect.bottom() - container_rect.bottom();
+            container.set_scroll_top((scroll_top + delta) as i32);
+        }
+    } else if active_rect.bottom() > container_rect.bottom() {
+        // No next-row anchor (last row) — fall back to bottom-aligning active.
+        let delta = active_rect.bottom() - container_rect.bottom();
+        container.set_scroll_top((scroll_top + delta) as i32);
+    }
+}
+
+/// Scrolls the `.operator__slides` container to its top. Used when the
+/// operator opens a new presentation so the first slide is visible without
+/// manual scroll-up. Issue #271 concern 3.
+pub(super) fn scroll_slides_to_top() {
+    let Some(document) = web_sys::window().and_then(|w| w.document()) else {
+        return;
+    };
+    let Ok(Some(container_el)) = document.query_selector(".operator__slides") else {
+        return;
+    };
+    let Ok(container) = container_el.dyn_into::<web_sys::HtmlElement>() else {
+        return;
+    };
+    container.set_scroll_top(0);
+}
+
+/// Returns the pixel distance one wheel notch should scroll the
+/// `.operator__slides` container. Measures the first rendered slide card's
+/// height + the grid row gap so the step adapts to user font-size scaling.
+/// Falls back to `DEFAULT_WHEEL_STEP_PX` if no card is rendered.
+fn step_for_wheel(container: &web_sys::HtmlElement) -> f64 {
+    let Ok(Some(card_el)) = container.query_selector(".operator__slide-card") else {
+        return DEFAULT_WHEEL_STEP_PX;
+    };
+    let Ok(card) = card_el.dyn_into::<web_sys::HtmlElement>() else {
+        return DEFAULT_WHEEL_STEP_PX;
+    };
+    let card_height = card.get_bounding_client_rect().height();
+    if card_height <= 0.0 {
+        return DEFAULT_WHEEL_STEP_PX;
+    }
+    // Grid row gap from operator.css `.operator__slides`: `gap: 0.9rem`.
+    // 0.9rem at 16px base = 14.4px. Hardcoded — if CSS changes, update here.
+    card_height + 14.4
+}
+
+/// Wheel handler for `.operator__slides`: intercepts the native (accelerated)
+/// scroll, applies a deterministic per-notch step instead. Issue #271 concern 2:
+/// neutralises macOS scroll acceleration so each notch advances ~1 row.
+pub(super) fn handle_wheel_event(ev: web_sys::WheelEvent) {
+    ev.prevent_default();
+    let direction = ev.delta_y().signum();
+    if direction == 0.0 {
+        return;
+    }
+    let Some(target) = ev.target() else { return };
+    let Ok(el) = target.dyn_into::<web_sys::Element>() else {
+        return;
+    };
+    let Ok(Some(container_el)) = el.closest(".operator__slides") else {
+        return;
+    };
+    let Ok(container) = container_el.dyn_into::<web_sys::HtmlElement>() else {
+        return;
+    };
+    let step = step_for_wheel(&container);
+    container.set_scroll_top((container.scroll_top() as f64 + direction * step) as i32);
+}
